Fix: I implemented hiding your API in env. You got undefined because i have rendered api in client side so because of that you cant store your api key secretely. So I moved the content to a new component named Home and rendered that component as client and passed the api key from process.env as argument in page.tsx
