+++ app/components/Input.tsx
@@ -10,6 +10,8 @@ const Input = ({ handleSearch, setLocation }: InputProps) => {
   return (
     <form className="flex items-center md:w-2/4 w-full order-2 md:order-1">
       <input
+        id="CityName"
+        name="CityName"
         type="text"
         className="w-full rounded-sm p-1 outline-none text-black"
         placeholder="Search..."
